Prioritizing Covid Tests
By this model we can efficiently reduce the covid cases by prioritizing the tests of some peoples. The data is randomly generated so do not believe in this answer. But if we can get the correct data then we can accurately do the task. This data depends on Runny Nose, Body Pain, Fever and Difficulty in breathing.